Move type-checker into its own module
