+++ src/type-checker/adapters/console-logger.ts
