+++ src/type-checker/index.ts
@@ -1,0 +1,4 @@
+export * from "./type-checker";
+export * from "./logger";
+export * from "./adapters/console-logger";
+export * from "./adapters/noop-logger";
